% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-labour.R
\docType{data}
\name{a1990_a}
\alias{a1990_a}
\title{Angrist (1990) Vietnam draft lottery, FICA earnings cells 1964-1977}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20440 rows and 12 columns.
}
\source{
\url{https://economics.mit.edu/people/faculty/josh-angrist/angrist-data-archive}
}
\usage{
a1990_a
}
\description{
Angrist (1990) Vietnam draft lottery, FICA earnings cells 1964-1977
}
\details{
Grouped Social Security earnings cells by birth cohort
(\code{byr}, 1944-1953), \code{race}, draft-lottery-number \code{interval} (73 bins of
5 numbers) and calendar \code{year} (1964-1977): mean/sd/n of FICA-taxable
earnings per cell. Column names as in the source file (\code{vmn1} mean
earnings, \code{vsd1} sd, \code{vnu1} n, \code{vfin1} fraction with zero earnings).
The 1978-1984 extension (also FICA and total W-2 pay) is \link{a1990_b}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("a1990")}
}
\keyword{datasets}
