% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-labour.R
\docType{data}
\name{aai2002}
\alias{aai2002}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("aai2002")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11204 rows and 19 columns.
}
\source{
\url{https://economics.mit.edu/people/faculty/josh-angrist/angrist-data-archive}
}
\usage{
aai2002
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("aai2002")}
}
\details{
11,204 JTPA (Job Training Partnership Act) applicants randomly
assigned to a training offer (\code{assignment}) with actual enrollment
(\code{training}) and total earnings in the following 30 months
(\code{earnings}); used for the paper's instrumental-variables quantile
treatment effects. \code{hsorged}, \code{married} and \code{wkless13} are hot-deck
imputed and can be fractional. Column dictionary: \code{subset(variables, dataset == "aai2002")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("aai2002")}
}
\keyword{datasets}
