% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bankruptcy.r
\docType{data}
\name{ABI_ch13}
\alias{ABI_ch13}
\title{Number of chapter 13 bankruptcy filings by State}
\format{Data frame with one column per state. Data are counts of chapter 13 bankruptcy cases.}
\source{
\url{http://www.abi.org/newsroom/bankruptcy-statistics}, but syndicated via the Federal Reserve System, so more data than contained in the public website.
}
\usage{
data(ABI_ch13)
}
\description{
Annual dataset of the number of chapter 13 bankruptcy filings for 1991-2016 by US state.
}
\author{
Florian Oswald
}
\keyword{datasets}

