% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bankruptcy.r
\docType{data}
\name{ABI_tot}
\alias{ABI_tot}
\title{Total US non-business bankruptcy filings by State}
\format{Data frame with one column per state. Data are counts of bankruptcy cases.}
\source{
\url{http://www.abi.org/newsroom/bankruptcy-statistics}, but syndicated via the Federal Reserve System, so more data than contained in the public website.
}
\usage{
data(ABI_tot)
}
\description{
Annual dataset of the number of non-business bankruptcy filings for 1960-2016 by US state.
}
\author{
Florian Oswald
}
\keyword{datasets}

