% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-labour.R
\docType{data}
\name{alo2009}
\alias{alo2009}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("alo2009")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1656 rows and 48 columns.
}
\source{
\url{https://economics.mit.edu/people/faculty/josh-angrist/angrist-data-archive}
}
\usage{
alo2009
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("alo2009")}
}
\details{
1,656 first-year students at a Canadian university randomized
into academic-support services and/or grade-based financial incentives
(STAR); grades, credits earned, probation status and survey responses.
Column dictionary from the archive's readme: \code{subset(variables, dataset == "alo2009")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("alo2009")}
}
\keyword{datasets}
