% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-monetary.R
\docType{data}
\name{bh2019}
\alias{bh2019}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("bh2019")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 614 rows and 5 columns.
}
\usage{
bh2019
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bh2019")}
}
\details{
Monthly 1975-02 to 2026-03 posterior-median structural oil
shocks from a Bayesian SVAR with incomplete sign/narrative
identification: \code{supply_shock}, \code{activity_shock} (global economic
activity), \code{consumption_demand_shock} and \code{inventory_demand_shock}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("bh2019")}
}
\keyword{datasets}
