% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{buildStateDistance}
\alias{buildStateDistance}
\title{State distance}
\usage{
buildStateDistance(
  miles = 100,
  output_path = "~/github/EconData/DataRepo/Miscellaneous/"
)
}
\arguments{
\item{miles}{(integer)}

\item{output_path}{(character)}
}
\description{
State distance
}
