% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-finance.R
\docType{data}
\name{bw2006}
\alias{bw2006}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("bw2006")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 792 rows and 15 columns.
}
\source{
\url{https://pages.stern.nyu.edu/~jwurgler/}
}
\usage{
bw2006
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bw2006")}
}
\details{
Monthly US investor sentiment 1958-01 to 2023-12: the sentiment
index (\code{SENT}) and its orthogonalized version (\code{SENT_ORTH}), the five
underlying proxies (\code{pdnd}, \code{ripo}, \code{nipo}, \code{cefd}, \code{s}) and macro
controls used to orthogonalize them. Column definitions:
\code{subset(variables, dataset == "bw2006")}. NYSE turnover was dropped as
a sixth sentiment proxy in this maintained version (unlike the original
2006/2007 papers). Snapshot updated 2024-05-31.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("bw2006")}
}
\keyword{datasets}
