% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bw2016}
\alias{bw2016}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("bw2016")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 266 rows and 36 columns.
}
\source{
\url{https://www.princeton.edu/~mwatson/ddisk/}
}
\usage{
bw2016
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bw2016")}
}
\details{
Quarterly 1947-2013 panel used to compare economic performance
under Democratic and Republican presidents. FRED mnemonics are kept as
column names; \code{subset(variables, dataset == "bw2016")} has the
descriptions.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("bw2016")}
}
\keyword{datasets}
