% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{chetty2016}
\alias{chetty2016}
\title{Chetty et al. (2016) life expectancy by household income percentile}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 9 columns.
}
\source{
\url{https://healthinequality.org/data/}
}
\usage{
chetty2016
}
\description{
Chetty et al. (2016) life expectancy by household income percentile
}
\details{
National life expectancy at age 40 by household income
percentile (1-100) and sex, 2001-2014, raw (\code{le_agg}) and
race-adjusted (\code{le_raceadj}) with standard errors and cell counts.
Released by the authors under a CC0 licence (public domain); cite the
2016 JAMA paper. State-level breakdown in \link{chetty2016_state}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("chetty2016")}
}
\seealso{
\link{chetty2016_state}
}
\keyword{datasets}
