% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{chetty2016_state}
\alias{chetty2016_state}
\title{Chetty et al. (2016) life expectancy by income quartile and state}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 43 columns.
}
\source{
\url{https://healthinequality.org/data/}
}
\usage{
chetty2016_state
}
\description{
Chetty et al. (2016) life expectancy by income quartile and state
}
\details{
Life expectancy at age 40 by household income quartile
(\code{le_agg_q1_F}...\code{le_agg_q4_M}, raw and race-adjusted, by sex), US
state, 2001-2014, with standard errors and cell counts. Same CC0
licence as \link{chetty2016}, which has the national breakdown.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("chetty2016")}
}
\seealso{
\link{chetty2016}
}
\keyword{datasets}
