% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-labour.R
\docType{data}
\name{ck1994}
\alias{ck1994}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("ck1994")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 410 rows and 46 columns.
}
\source{
\url{https://davidcard.berkeley.edu/data_sets.html}
}
\usage{
ck1994
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("ck1994")}
}
\details{
410 New Jersey and Pennsylvania fast-food restaurants surveyed
just before and about 8 months after New Jersey's April 1992 minimum-wage
increase (the \verb{_2} suffix marks wave-2 columns); \code{state} is 1 for New
Jersey. The canonical difference-in-differences dataset. Column
dictionary from the source codebook: \code{subset(variables, dataset == "ck1994")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("ck1994")}
}
\keyword{datasets}
