% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_coord_divisions.r
\docType{data}
\name{coordDivision}
\alias{coordDivision}
\title{Average coordinates of census division center}
\format{data.table}
\source{
\url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\usage{
data(coordDivision)
}
\description{
arithmethic average of population centroids
of US states, grouped by census division.
}
\author{
Florian Oswald
}

