\docType{data}
\name{coordStates}
\alias{coordStates}
\title{Coordinates of US State centers}
\source{
  \url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\description{
  data.frame with lat long coordinates of geographic center
  of a US state
}
\author{
  Florian Oswald
}
\keyword{datasets}

