\docType{data}
\name{CPIAUCSL}
\alias{CPIAUCSL}
\title{CPI all urban consumers seasonally adjusted}
\format{monthly xts series}
\source{
  \url{http://research.stlouisfed.org/fred2/series/CPIAUCSL}
}
\description{
  monthly CPI data for all urban consumers, seasonally
  adjusted. index 1982-84=100.
}
\author{
  Florian Oswald
}
\keyword{datasets}

