\docType{data}
\name{CPIHOSSL}
\alias{CPIHOSSL}
\title{CPI all urban consumers (Housing)}
\format{monthly xts series}
\source{
  \url{http://research.stlouisfed.org/fred2/series/CPIHOSSL}
}
\description{
  monthly CPI data for all urban consumers, housing,
  seasonally adjusted. index 1982-84=100.
}
\author{
  Florian Oswald
}
\keyword{datasets}

