% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_distance-division.r
\docType{data}
\name{Division_distMat}
\alias{Division_distMat}
\title{Haversine Distance Matrix between center of US census Divisions}
\format{numeric matrix}
\source{
\url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\usage{
data(Division_distMat)
}
\description{
contains a matrix with great-circle distance
in km between the center of US census Divisions.
}
\author{
Florian Oswald
}
\references{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

