% Generated by roxygen2 (4.0.0): do not edit by hand
\name{doGetPath}
\alias{doGetPath}
\title{Generates paths to resample coefficients or models in String format}
\usage{
doGetPath(prefix, modelType, countryAbbrev, energyType = "iQ",
  factor = "iK", baseResample)
}
\arguments{
  \item{prefix}{the prefix for the model file names.}

  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ".}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}
}
\value{
the relative path of the file containing the data for the requested resample data.
}
\description{
This function returns a string representing the relative file path for the resample coefficients or models being requested.
}

