% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCBP.R
\name{downloadCBP}
\alias{downloadCBP}
\title{Download CBP data}
\usage{
downloadCBP(years = 2017, location = "national", output_path)
}
\arguments{
\item{years}{(integer) any integer between 2000 and 2017 is supported.}

\item{location}{(character) options are "county", "state", "national".}
}
\description{
Download CBP data
}
