% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-monetary.R
\docType{data}
\name{eampd_pc}
\alias{eampd_pc}
\title{Altavilla et al. (2019) EA-MPD, press conference window}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 315 rows and 46 columns.
}
\usage{
eampd_pc
}
\description{
Same columns as \link{eampd}, restricted to the press-conference Q&A window;
\code{NA} on dates with no press conference (Governing Council statement only).
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("eampd")}
}
\seealso{
\link{eampd}, \link{eampd_pr}
}
\keyword{datasets}
