% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-monetary.R
\docType{data}
\name{eampd_pr}
\alias{eampd_pr}
\title{Altavilla et al. (2019) EA-MPD, press release window}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 315 rows and 46 columns.
}
\usage{
eampd_pr
}
\description{
Same columns as \link{eampd}, restricted to the press-release window (the
initial Governing Council statement, narrower than the combined event
window).
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("eampd")}
}
\seealso{
\link{eampd}, \link{eampd_pc}
}
\keyword{datasets}
