% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econdata-package.R
\docType{package}
\name{econdata}
\alias{econdata-package}
\alias{econdata}
\title{econdata: Curated List of Economic Datasets}
\description{
A curated collection of datasets used in influential empirical macroeconomics papers, with a catalog of categories and tags, pointers to external replication data, and CSV copies served from the package website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvasilopoulos/econdata}
  \item Report bugs at \url{https://github.com/kvasilopoulos/econdata/issues}
}

}
\author{
\strong{Maintainer}: Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}

Authors:
\itemize{
  \item Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}
}

}
\keyword{internal}
