% Generated by roxygen2 (4.0.0): do not edit by hand
\name{energyTypeFromFormula}
\alias{energyTypeFromFormula}
\title{Extracts an energyType from a formula}
\usage{
energyTypeFromFormula(formulaStr, formula = eval(parse(text = formulaStr)))
}
\arguments{
  \item{formulaStr}{a string representing the formula that
  was fitted}

  \item{formula}{the formula that was fitted}
}
\value{
a string representing the energyType that was used for the fitting.
NA if the sfModel was used.
}
\description{
Extracts an energyType from a formula
}

