% Generated by roxygen2 (4.0.2): do not edit by hand
\name{extractEnergyType}
\alias{extractEnergyType}
\title{Extracts an energyType from a formula or a factorString.}
\usage{
extractEnergyType(x, eTypes = energyTypes, sep = "+")
}
\arguments{
\item{x}{a formula or factorString from which you want to extract an energyType}

\item{sep}{the separator for the formula or factorString}

\item{energyTypes}{the types of energy available}
}
\value{
a string representing the energyType that was used for the fitting. NA if no energy type was found.
}
\description{
Extracts an energyType from a formula or a factorString.
}

