% Generated by roxygen2 (4.0.0): do not edit by hand
\name{factorFromFormula}
\alias{factorFromFormula}
\title{Extracts a factor from a formula}
\usage{
factorFromFormula(formulaStr, formula = eval(parse(text = formulaStr)))
}
\arguments{
  \item{formulaStr}{a string representing the formula that
  was fitted}

  \item{formula}{the formula that was fitted}
}
\value{
a string representing the factor that was used for the fitting.
NA if the singleFactorModel was not used.
}
\description{
Extracts a factor from a formula
}

