% Generated by roxygen2 (4.0.0): do not edit by hand
\name{factorsFromFormula}
\alias{factorsFromFormula}
\title{Extracts factors of production from a formula}
\usage{
factorsFromFormula(formula)
}
\arguments{
  \item{formula}{the formula that was fitted}
}
\value{
a string representing the factors that were used for the fitting.
NA if the sfModel was not used.
}
\description{
Extracts factors of production from a formula
}

