% Generated by roxygen2 (4.0.0): do not edit by hand
\name{factorString}
\alias{factorString}
\title{Creates an string representation of the factors of production}
\usage{
factorString(formula, nest, sep = "+", showNestParens = FALSE)
}
\arguments{
  \item{formula}{the formula used for fitting. The energy
  variable is extracted from the formula.}

  \item{nest}{an integer vector containing 2 or 3 values.
  For 2-value vectors, integers must be 1 or 2. For 3-value
  vectors, integers must be 1, 2, or 3.}

  \item{sep}{the separator for the factorString}

  \item{showNestParens}{tells whether to show parentheses
  in the factor string if a nest is present}
}
\value{
a string representing the nesting of the form \code{iQ+iK+iL} or \code{(iQ + iK) + (iL)}, etc.
An empty string if a nest is not involved.
}
\description{
Creates an string representation of the factors of production
}
\details{
Factors of production are ordered according to the nest, if present.
In the nest argument, the integer 1 indicates the capital stock variable ("iK"),
the integer 2 indicates the labor variable ("iL"), and
the integer 3 indicates the energy variable (one of "iQ", "iX", or "iU")
Nesting positions are given by location in the \code{nest} vector, if present.
c(3,1,2) is interpreted as energy and capital stock nested together.
For example, (iX + iK) + (iL).
If the \code{nest} vector is not present, factors of production
are returned in the order they appear in the \code{formula}.
If showNestParens=FALSE, parentheses are not included in the factor string, e.g., \code{iX+iK+iL}.
If showNestParens=TRUE and a nest argument is provided,
parentheses are include in the factor string, e.g., \code{(iX + iK) + (iL)} or \code{(iL) + ()}.
}

