% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bankruptcy.r
\docType{data}
\name{Fed_pop_count}
\alias{Fed_pop_count}
\title{US State Population Count}
\format{Data frame with one column per state. Data are the number of people residing in a state.}
\source{
Syndicated via the Federal Reserve System.
}
\usage{
data(Fed_pop_count)
}
\description{
Annual data on population by US state
}
\author{
Florian Oswald
}
\keyword{datasets}

