\docType{data}
\name{FHFA_Div}
\alias{FHFA_Div}
\title{FHFA house price indices by US census Division}
\format{long}
\source{
  \url{http://www.fhfa.gov/Default.aspx?Page=87},
  \url{http://www.fhfa.gov/webfiles/26093/4q13hpicd_expandeddata.txt}
}
\description{
  contains 2 lists \code{qtr} and \code{yr}, which stands
  for yearly and quarterly frequency. all data is in long
  format as data.tables.
}
\author{
  Florian Oswald
}
\keyword{datasets}

