\docType{data}
\name{FHFA_msa50}
\alias{FHFA_msa50}
\title{FHFA house price indices for 50 biggest MSAs over time}
\format{long}
\source{
  \url{http://www.fhfa.gov/Default.aspx?Page=87}
}
\description{
  contains 2 lists \code{qtr} and \code{yr}, which stands
  for yearly and quarterly frequency. all data is in long
  format as data.tables.
}
\author{
  Florian Oswald
}
\keyword{datasets}

