% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_FHFA_states.r
\docType{data}
\name{FHFA_states}
\alias{FHFA_states}
\title{FHFA house price indices US states over time}
\format{long}
\source{
\url{http://www.fhfa.gov/Default.aspx?Page=87}, \url{http://www.fhfa.gov/webfiles/26093/4q13hpicd_expandeddata.txt}
}
\usage{
data(FHFA_states)
}
\description{
contains 2 lists \code{qtr} and \code{yr},
which stands for yearly and quarterly frequency.
all data is in long format as data.tables.
}
\author{
Florian Oswald
}

