% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fittingID}
\alias{fittingID}
\title{Creates an id for this run of resampling}
\usage{
fittingID(Source, countryAbbrev, fitfun, formula, nest = NULL, n,
  sep = " : ")
}
\arguments{
\item{Source}{the data source used for this fit}

\item{countryAbbrev}{the country being fitted}

\item{fitfun}{the function used for fitting}

\item{formula}{the formula used for the fitting}

\item{nest}{if used, the nest employed for this fit. A 2- or 3-vector of integers.}

\item{n}{the number of resamples being attempted}

\item{sep}{the separator used to create the id string. Default is " : ".}
}
\value{
a string to be used as the id for this resample
}
\description{
Creates an id for this run of resampling
}

