% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{gcd.hf}
\alias{gcd.hf}
\title{Haversine Distance function}
\usage{
gcd.hf(long1, lat1, long2, lat2)
}
\description{
use the haversine formula to compute distance
between 2 locations
}
\references{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

