% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCBP.R
\name{getCBP}
\alias{getCBP}
\title{Prepare CBP data}
\usage{
getCBP(
  years = 2017,
  location = "national",
  industry = 0,
  LFO = "-",
  input_path,
  output_path
)
}
\arguments{
\item{years}{(integer) any integer between 2000 and 2017 is supported.}

\item{location}{(character) options are "county", "state", "national".}

\item{industry}{(integer) options are 0, 2, 3, 4, 6.}

\item{LFO}{(character) legal form of organization.}
}
\description{
Prepare CBP data
}
