% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorpTax.R
\name{getCorpTaxSources}
\alias{getCorpTaxSources}
\title{Download corporate tax rate data from various sources}
\usage{
getCorpTaxSources(source_path)
}
\arguments{
\item{source_path}{(character)}
}
\description{
Download corporate tax rate data from various sources
}
