% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getCPI}
\alias{getCPI}
\title{Download and prepare FRED CPI data (measured on January 1)}
\usage{
getCPI(output_path = "~/github/EconData/DataRepo/Miscellaneous/")
}
\arguments{
\item{output_path}{(character)}
}
\description{
Download and prepare FRED CPI data (measured on January 1)
}
