% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getCZ}
\alias{getCZ}
\title{Download and prepare commuting zone data}
\usage{
getCZ(output_path = "~/github/EconData/DataRepo/Miscellaneous/")
}
\arguments{
\item{output_path}{(character)}
}
\description{
Download and prepare commuting zone data
}
