% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getDistances}
\alias{getDistances}
\title{Get all distances}
\usage{
getDistances(
  miles = 100,
  output_path = "~/github/EconData/DataRepo/Miscellaneous/"
)
}
\arguments{
\item{miles}{(integer)}

\item{output_path}{(character)}
}
\description{
Get all distances
}
