% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getGDP}
\alias{getGDP}
\title{Download and prepare BEA GDP data, collapsed by CZ and NAICS-2}
\usage{
getGDP(output_path = "~/github/EconData/DataRepo/Miscellaneous/")
}
\arguments{
\item{output_path}{(character)}
}
\description{
Download and prepare BEA GDP data, collapsed by CZ and NAICS-2
}
