% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorpTax.R
\name{getStateCorpTax}
\alias{getStateCorpTax}
\title{Download and manually correct corporate tax rate data}
\usage{
getStateCorpTax(source_file)
}
\arguments{
\item{source_file}{(character)}
}
\description{
Download and manually correct corporate tax rate data
}
