% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getStateFips}
\alias{getStateFips}
\title{Download and prepare state FIPS codes}
\usage{
getStateFips(output_path = "~/github/EconData/DataRepo/Miscellaneous/")
}
\arguments{
\item{output_path}{(character)}
}
\description{
Download and prepare state FIPS codes
}
