% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMisc.R
\name{getTradables}
\alias{getTradables}
\title{Download and prepare tradables data}
\usage{
getTradables(output_path = "~/github/EconData/DataRepo/Miscellaneous/")
}
\arguments{
\item{output_path}{(character)}
}
\description{
Download and prepare tradables data
}
