% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{getUS_inflation}
\alias{getUS_inflation}
\title{get US inflation indexed to a base year}
\usage{
getUS_inflation(idx = "2012 Q1", freq = "quarterly")
}
\description{
returns US CPI from http://research.stlouisfed.org/fred2/series/CPIAUCSL indexed at idx and at frequency freq = "monthly", "quartrely" or "yearly".
}
\examples{
getUS_inflation(idx="1996-01-01",freq="monthly")
getUS_inflation(idx="1996 Q3",freq="quarterly")
getUS_inflation(idx="1996",freq="yearly")
}

