% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gk2015}
\alias{gk2015}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("gk2015")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 396 rows and 13 columns.
}
\usage{
gk2015
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("gk2015")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("gk2015")}
}
\keyword{datasets}
