% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gss2005}
\alias{gss2005}
\title{Gurkaynak, Sack and Swanson (2005) target and path factors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 212 rows and 3 columns.
}
\source{
\url{http://www.bilkent.edu.tr/~refet/}
}
\usage{
gss2005
}
\description{
Gurkaynak, Sack and Swanson (2005) target and path factors
}
\details{
One row per FOMC announcement 1994-2018 with the target and
path factors, as updated by Gurkaynak, Karasoy-Can and Lee (2021,
Journal of Finance). The raw asset-price surprises they are built from
are in \link{gss2005_surprises}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("gss2005")}
}
\keyword{datasets}
