% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gz2012}
\alias{gz2012}
\title{Gilchrist and Zakrajsek (2012) excess bond premium}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 643 rows and 4 columns.
}
\source{
\url{https://www.federalreserve.gov/econres/notes/feds-notes/updating-the-recession-risk-and-the-excess-bond-premium-20161006.html}
}
\usage{
gz2012
}
\description{
Gilchrist and Zakrajsek (2012) excess bond premium
}
\details{
Monthly GZ credit spread (\code{gz_spread}), excess bond premium
(\code{ebp}) and the implied 12-month-ahead recession probability
(\code{est_prob}), as updated by the Federal Reserve Board. Snapshot of
September 2026.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("gz2012")}
}
\keyword{datasets}
