% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-finance.R
\docType{data}
\name{hkm2017_monthly}
\alias{hkm2017_monthly}
\title{He, Kelly and Manela (2017) intermediary capital ratio and risk factor (monthly)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 666 rows and 5 columns.
}
\source{
\url{https://zhiguohe.net/data-and-empirical-patterns/}
}
\usage{
hkm2017_monthly
}
\description{
He, Kelly and Manela (2017) intermediary capital ratio and risk factor (monthly)
}
\details{
Monthly counterpart of \link{hkm2017_quarterly}, 1970-01-. Same four
columns. Snapshot of the 2025-06-27 update.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("hkm2017")}
}
\keyword{datasets}
