% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-finance.R
\docType{data}
\name{hkm2017_quarterly}
\alias{hkm2017_quarterly}
\title{He, Kelly and Manela (2017) intermediary capital ratio and risk factor (quarterly)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 222 rows and 5 columns.
}
\source{
\url{https://zhiguohe.net/data-and-empirical-patterns/}
}
\usage{
hkm2017_quarterly
}
\description{
He, Kelly and Manela (2017) intermediary capital ratio and risk factor (quarterly)
}
\details{
Quarterly primary-dealer intermediary capital ratio
(\code{intermediary_capital_ratio}), capital risk factor
(\code{intermediary_capital_risk_factor}), value-weighted investment return
and squared leverage ratio, 1970Q1-. Maintained by Zhiguo He; snapshot
of the 2025-06-27 update. Monthly version: \link{hkm2017_monthly}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("hkm2017")}
}
\keyword{datasets}
