% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_landvalues.r
\docType{data}
\name{HomeValues}
\alias{HomeValues}
\title{Home, Structure and Land Values by state over time}
\format{quarterly}
\source{
\url{https://www.lincolninst.edu/subcenters/land-values/}
}
\usage{
data(HomeValues)
}
\description{
quarterly data by state on 
\itemize{
\item mean home value (current dollars),
\item mean structure cost (current dollars),
\item mean land value (current dollars),
\item land share,
\item home price index,
\item land price index,
}
The data have been compiled by Morris A. Davis, U of Wisconsin-Madison
}
\author{
Florian Oswald
}
\references{
\url{http://morris.marginalq.com/}, 
\url{https://www.lincolninst.edu/subcenters/land-values/}
}

