% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-updated.R
\docType{data}
\name{hprice}
\alias{hprice}
\title{Historical housing market}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 130 rows and 5 columns.}
\usage{
hprice
}
\description{
asdfasdf
}
\keyword{datasets}
