% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hprice_iho}
\alias{hprice_iho}
\title{International House Price Database}
\format{quarterly data from 1975-Q1 to 2018-Q4}
\source{
https://www.dallasfed.org/institute/houseprice#tab2
}
\usage{
hprice_iho
}
\description{
\itemize{
}
}
\keyword{datasets}
