% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{j2002}
\alias{j2002}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("j2002")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 9 columns.
}
\source{
\url{https://web.stanford.edu/~chadj/}
}
\usage{
j2002
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("j2002")}
}
\details{
US annual growth-accounting series 1950-1993: real GDP, hours,
capital stock, GDP per hour (\code{gdp_l}), capital-output ratio (\code{k_y}),
saving rate (\code{sk}), average educational attainment (\code{edatt}) and total
factor productivity (\code{a}). The scientists/engineers and employment
tables for France, Germany, Japan, the UK and the G-5 total in the same
source file are not included.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("j2002")}
}
\keyword{datasets}
