% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jk2020}
\alias{jk2020}
\title{Jarocinski and Karadi (2020) Fed monetary policy and information shocks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 353 rows and 7 columns.
}
\source{
\url{https://marekjarocinski.github.io/jkshocks/jkshocks.html}
}
\usage{
jk2020
}
\description{
Jarocinski and Karadi (2020) Fed monetary policy and information shocks
}
\details{
Monthly sums of FOMC-announcement surprises 1990-2019: the
policy-indicator principal component (\code{pc1ff1_hf}) and the S&P 500
(\code{sp500_hf}), decomposed into monetary policy (\code{mp_}) and central bank
information (\code{cbi_}) shocks with poor man's (\verb{_pm}) and median
(\verb{_median}) sign restrictions.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("jk2020")}
}
\keyword{datasets}
