% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jk2020_ecb}
\alias{jk2020_ecb}
\title{Jarocinski and Karadi (2020) ECB monetary policy and information shocks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 294 rows and 7 columns.
}
\source{
\url{https://marekjarocinski.github.io/jkshocks/jkshocks.html}
}
\usage{
jk2020_ecb
}
\description{
Jarocinski and Karadi (2020) ECB monetary policy and information shocks
}
\details{
ECB counterpart of \link{jk2020}: monthly sums of surprises around
ECB announcements 1999-2023 from the Altavilla et al. (2019) Euro Area
Monetary Policy event-study Database, split into monetary policy
(\code{mp_}) and central bank information (\code{cbi_}) shocks.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("jk2020")}
}
\keyword{datasets}
