% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{jln2015}
\alias{jln2015}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("jln2015")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 792 rows and 10 columns.
}
\source{
\url{https://www.sydneyludvigson.com/macro-and-financial-uncertainty-indexes}
}
\usage{
jln2015
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("jln2015")}
}
\details{
Monthly macro, financial and real uncertainty indexes at 1, 3
and 12 month horizons (\code{macro_h1}, ..., \code{real_h12}). August 2026 update.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("jln2015")}
}
\keyword{datasets}
