% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadAllResampleData}
\alias{loadAllResampleData}
\title{Loads all resample data, for the base fit and resample fits}
\usage{
loadAllResampleData(modelType, energyType = "none", factor,
  countryAbbrevs = countryAbbrevs, archive = NULL, baseResample = NULL)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ". If
  the energyType argument is missing or NA, you'll get data
  for the CES model without energy.}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{countryAbbrevs}{a vector of country abbreviations
  that you want in the resulting data frame}

  \item{archive}{path to a .zip archive containing resample
  model data, if resample data are to be loaded from the
  archive}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}
}
\value{
a data frame containing resample data for specified countries for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function returns a data frame that contains all of the resample model coefficients
}

