% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadCESResampleData}
\alias{loadCESResampleData}
\title{Loads and binds data for a CES resample ternary plot.}
\usage{
loadCESResampleData(nest, energyType = "none", archive = NULL, baseResample)
}
\arguments{
  \item{nest}{the desired nest. One of \code{"(kl)"},
  \code{"(kl)e"}, \code{"(le)k"}, or \code{"(ek)l"}. If you
  specify nest=\code{"all"}, you'll get data for all nests.
  You'll need to specify energyType if you use
  nest=\code{"all"}.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ". If
  the energyType argument is missing or NA, you'll get data
  for the CES model without energy.}

  \item{archive}{path to a zip file archive}

  \item{baseResample}{the parent directory of all resample
  data, usually \code{"data_resample"}.}
}
\value{
a data frame containing CES resample data for the given arguments.
}
\description{
Loads and binds data for a CES resample ternary plot.
}

