% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadCESSpaghettiGraphData}
\alias{loadCESSpaghettiGraphData}
\title{Creates a data frame containing historical data, the fit to historical data, and resample predictions.}
\usage{
loadCESSpaghettiGraphData(nest = "(kl)", energyType = "none",
  archive = NULL, baseHistorical, baseResample)
}
\arguments{
  \item{nest}{the desired nest. One of \code{"(kl)"},
  \code{"(kl)e"}, \code{"(le)k"}, or \code{"(ek)l"}. If you
  specify nest=\code{"all"}, you'll get data for all nests.
  You'll need to specify energyType if you use
  nest=\code{"all"}.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ". If
  the energyType argument is missing or NA, you'll get data
  for the CES model without energy.}

  \item{archive}{path to a zip file archive}

  \item{baseResample}{the parent directory of all resample
  data, usually \code{"data_resample"}.}
}
\value{
a data frame containing CES data for a "spaghetti" graph.
}
\description{
Creates a data frame containing historical data, the fit to historical data, and resample predictions.
}

