% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadResampleDataBaseFitOnly}
\alias{loadResampleDataBaseFitOnly}
\title{Loads resample data, for the base fit only (ignoring the resamples)}
\usage{
loadResampleDataBaseFitOnly(modelType, countryAbbrev, energyType = "none",
  factor = "K", archive = NULL, baseResample)
}
\arguments{
  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev}{the country for which you want to
  load resample data}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ". If
  the energyType argument is missing or NA, you'll get data
  for the CES model without energy.}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{archive}{path to a .zip archive containing resample
  model data, if resample data are to be loaded from the
  archive}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data, if the data are
  to be loaded from the directory.}
}
\value{
coefficients for the base fit (not the resample fits) for the specified country
for the given \code{modelType}, \code{energyType},
or \code{factor}.
}
\description{
This function returns a data frame that contains base fit coefficients
}

