\name{makeAllData}
\alias{makeAllData}
\title{make all datasets}
\usage{
  makeAllData(root = "~/git/EconData")
}
\arguments{
  \item{root}{path/to/your/package/root}
}
\description{
  make all datasets
}
\seealso{
  Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeCPIUS}}, \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

