% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeBEAincome}
\alias{makeBEAincome}
\title{Get BEA annual personal income by state for 1949-now}
\usage{
makeBEAincome(root = "~/git/EconData")
}
\description{
data obtained from query to
\url{http://www.bea.gov/iTable/iTableHtml.cfm?reqid=70&step=30&isuri=1&7022=36&7023=0&7033=-1&7024=non-industry&7025=0&7026=xx&7027=-1&7001=336&7028=10&7031=0&7040=-1&7083=levels&7029=36&7090=70}
}

