% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeCPIUS}
\alias{makeCPIUS}
\title{make several US inflation datasets}
\usage{
makeCPIUS()
}
\description{
download data from FRED with quantmod package
}
\details{
saves cpi for all urban consumers 
and housing cpi to disk
}
\references{
\url{http://research.stlouisfed.org/fred2/series/CPIAUCSL} 
            \url{http://research.stlouisfed.org/fred2/series/CPIHOSSL} 
            \url{http://research.stlouisfed.org/fred2/series/USSTHPI}
}
\seealso{
Other FREDs: \code{\link{makeMORTGAGE30US}}

Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMORTGAGE30US}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

