% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeFHFA}
\alias{makeFHFA}
\title{make FHFA expanded house price indices}
\usage{
makeFHFA()
}
\value{
NULL.
}
\description{
downloads and builds FHFA 
Expanded-Data Indexes (Estimated using 
Enterprise, FHA, and Real Property County Recorder 
Data Licensed from DataQuick). Data is quarterly.
}
\references{
\url{http://www.fhfa.gov/Default.aspx?Page=87}
}
\seealso{
Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMORTGAGE30US}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

