% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makers.r
\name{makeInterstateMig}
\alias{makeInterstateMig}
\title{make Interstate Migration Transition Matrix}
\usage{
makeInterstateMig(root = "~/git/EconData")
}
\arguments{
\item{root}{}
}
\description{
make Interstate Migration Transition Matrix
}

