% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeLincolnHomeValues}
\alias{makeLincolnHomeValues}
\title{make Lincoln Institute Home and Land Value Dataset}
\usage{
makeLincolnHomeValues()
}
\value{
NULL.
}
\description{
reads data from the Lincoln Institute xls file.
}
\references{
\url{https://www.lincolninst.edu/subcenters/land-values/}
}
\seealso{
Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeFHFA}}, \code{\link{makeMORTGAGE30US}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

