% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeMedianIncome}
\alias{makeMedianIncome}
\title{make dataset of median income by US state and year}
\usage{
makeMedianIncome()
}
\value{
NULL.
}
\description{
produce dataset in current and 2012 dollars of
median household income by US state.
}
\details{
saves 2 lists \code{current} and \code{in2012}, each 
with elements zinc, zse, incl, sel. Those elements are
\enumerate{
\item zinc: a zoo object of median income
\item zse: a zoo object of it's standard error
\item incl: data.frame with median income in long format
\item sel: data.frame with standard error of median income in long format
}

source: census bureau
}
\references{
\url{http://www.census.gov/hhes/www/income/data/historical/household/}
}
\seealso{
Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMORTGAGE30US}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

