% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makers.r
\name{makeOwnershipRates}
\alias{makeOwnershipRates}
\title{make Homeownership rates by state over time}
\usage{
makeOwnershipRates(root = "~/git/EconData")
}
\arguments{
\item{root}{}
}
\description{
make Homeownership rates by state over time
}

