% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makePopulation}
\alias{makePopulation}
\title{Census Population Count by state for 1948-2012}
\usage{
makePopulation()
}
\description{
takes data from US census bureau intercensal 
sites at
http://www.census.gov/popest/data/state/asrh/1980s/80s_st_totals.html
http://www.census.gov/popest/data/state/asrh/1980s/tables/8090com.txt
http://www.census.gov/popest/data/state/totals/1990s/tables/ST-99-03.txt
http://www.census.gov/popest/data/intercensal/state/tables/ST-EST00INT-01.xls
http://www.census.gov/popest/data/national/totals/2015/files/NST-EST2015-alldata.csv
where the last two have been manually cleaned in inst/extdata as \code{pop2000.csv} and \code{pop2010.csv}
http://www.census.gov/popest/data/intercensal/state/tables/ST-EST00INT-01.xls
http://www.census.gov/popest/data/state/totals/2012/tables/NST-EST2012-01.csv
}

