\name{makeUS_coordinates}
\alias{makeUS_coordinates}
\title{make Coordinate of US State centers}
\usage{
  makeUS_coordinates(root = "~/git/EconData")
}
\description{
  make Coordinate of US State centers
}
\references{
  \url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\seealso{
  Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_distance}}
}

