\name{makeUS_distance}
\alias{makeUS_distance}
\title{make Distance between US state centers}
\usage{
  makeUS_distance(root = "~/git/EconData")
}
\description{
  make Distance between US state centers
}
\seealso{
  Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}}
}

