% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_bankruptcy.r
\docType{data}
\name{MBA_fore_rate}
\alias{MBA_fore_rate}
\title{Rate of Foreclosure Starts}
\format{Data frame with one column per state.}
\source{
Mortgage Banker's Association syndicated via the Federal Reserve System.
}
\usage{
data(fore_rate)
}
\description{
"Foreclosures started" refers to the percentage rate of loans for which a foreclosure has been initiated during the quarter, that is, the number of loans sent to the foreclosure process as a percentage of the total number of mortgages in the pool.
}
\author{
Florian Oswald
}
\keyword{datasets}

