\docType{data}
\name{Migration}
\alias{Migration}
\title{Migratory Flows between US states in 2012}
\format{long data.table}
\source{
  \url{http://www.census.gov/hhes/migration/data/acs/state-to-state.html}
}
\description{
  Census ACS estimate of the flow of migrants from one US
  state to another for 2012.
}
\author{
  Florian Oswald
}
\keyword{datasets}

