% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mmo2018}
\alias{mmo2018}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("mmo2018")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67 rows and 57 columns.
}
\source{
\url{https://karelmertens.com/research/}
}
\usage{
mmo2018
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("mmo2018")}
}
\details{
Annual 1946-2012 average marginal tax rates by income group
(\verb{amtr_*} total, \verb{amiitr_*} income tax, \verb{amptr_*} payroll), narrative
marginal-rate shocks (\verb{shock_*}), log average income (\verb{linc_*}) and the
macro controls of the paper (\verb{ctrl_*}).
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("mmo2018")}
}
\keyword{datasets}
