% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mr2013}
\alias{mr2013}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("mr2013")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 228 rows and 22 columns.
}
\source{
\url{https://karelmertens.com/research/}
}
\usage{
mr2013
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("mr2013")}
}
\details{
Quarterly 1950-2006: narrative personal (\code{T_PI}) and corporate
(\code{T_CI}) income tax liability changes, their unanticipated parts
(\code{m_PI}, \code{m_CI}), average tax rates, tax bases and the macro
aggregates of the SVAR. Column names as in the replication files.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("mr2013")}
}
\keyword{datasets}
