% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mr2014}
\alias{mr2014}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("mr2014")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 228 rows and 9 columns.
}
\source{
\url{https://karelmertens.com/research/}
}
\usage{
mr2014
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("mr2014")}
}
\details{
Quarterly 1950-2006 log real tax revenues (\code{ltax}), government
spending (\code{lgov}) and output (\code{lgdp}), the Romer-Romer narrative tax
measure (\code{tax_narrative}) and its alternatives.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("mr2014")}
}
\keyword{datasets}
