% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/files.R
\name{natmetaFrame}
\alias{natmetaFrame}
\title{Natural coefficients and other metadata for all model attempts}
\usage{
natmetaFrame(object)
}
\arguments{
\item{object}{the model object from which you want to extract
information from all model attempts.}
}
\value{
a data frame containing the \code{naturalCoeffs} and other metadata from
all model attempts in \code{object}.
}
\description{
A convenience function that returns a data frame containing
\code{naturalCoeffs} and additional metadata for all model attempts
stored in the \code{model.attempts} attribute of \code{object}.
}

