% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nestFromFormula}
\alias{nestFromFormula}
\title{Extracts a nest string from a formula}
\usage{
nestFromFormula(formula, Kvar = "iK", Lvar = "iL")
}
\arguments{
  \item{formulaStr}{a string representing the formula that
  was fitted}

  \item{formula}{the formula that was fitted}
}
\value{
a string representing the nesting that was used.  NULL if we can't figure out the nest.
The form is \code{"(kl)e"} or similar.
This function assumes that we are using a CES production function.

It also assumes that the functional form is \code{response ~ fp + fp + fp + iYear},
where \code{fp} are factors of production.
}
\description{
Extracts a nest string from a formula
}

