% Generated by roxygen2 (4.0.0): do not edit by hand
\name{nestString}
\alias{nestString}
\title{Creates an string representation of the nesting}
\usage{
nestString(nest)
}
\arguments{
  \item{nest}{an integer vector containing 2 or 3 values.
  For 2-value vectors, integers must be 1 or 2. For 3-value
  vectors, integers must be 1, 2, or 3.}
}
\value{
a string representing the nesting of the form "(kl)e", etc.
An empty string if a nest is not involved.
}
\description{
Creates an string representation of the nesting
}

