% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-monetary.R
\docType{data}
\name{ns2018_1995}
\alias{ns2018_1995}
\title{Nakamura and Steinsson (2018) monetary policy news shock, extended sample}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 162 rows and 4 columns.
}
\usage{
ns2018_1995
}
\description{
Same construction as \link{ns2018} but on the larger Jan 1995-Mar 2014 sample:
includes unscheduled meetings (flagged by \code{unscheduled}) and the
financial-crisis window (still excludes the two Sep 2001 meetings after
9/11). \code{policy_news_shock} differs slightly from \link{ns2018} on overlapping
dates because the first principal component is re-estimated on this
sample.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("ns2018")}
}
\seealso{
\link{ns2018}
}
\keyword{datasets}
