% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{papers}
\alias{papers}
\title{List of the contained papers}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 19 columns.
}
\usage{
papers
}
\description{
One row per bibliography entry in \code{data-raw/bib/papers.bib}. \code{BIBTEXKEY}
matches the \code{paper} column of \link{catalog}.
}
\seealso{
\link{catalog}, \link{sources}, \link{variables}
}
\keyword{datasets}
