% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parseFactorString}
\alias{parseFactorString}
\title{Parse a factor string}
\usage{
parseFactorString(factorString, sep = "+", rVar = "iGDP",
  kVar = factors[["K"]], lVar = factors[["L"]], tVar = "iYear")
}
\arguments{
  \item{factorString}{a string of the form
  \code{"iK+iL+iU"}.}

  \item{sep}{the separator in \code{factorString}}

  \item{rVar}{the name of the response variable}

  \item{kVar}{the name of the capital stock variable}

  \item{lVar}{the name of the labor variable}

  \item{tVar}{the name of the time variable}
}
\value{
a list with two elements, \code{formula} and \code{nest}.
}
\description{
Parse a factor string
}
\details{
White space is removed from \code{factorString}.
In the return value, \code{formula} is a formula object with factors of production
in the correct order.
\code{nest} is an integer vector whose elements indicate nesting of the
factors of production. For example, \code{c(3,1,2)} means that the 3rd and 1st factors of production
in the formula are nested together and the 2nd factor of production
is independent.
}

