% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parseID}
\alias{parseID}
\title{Parses the id created by resampling}
\usage{
parseID(id, sep = " : ")
}
\arguments{
  \item{id}{an \code{id} as a string}

  \item{sep}{the separator used to create the id string.
  Default is " : ".}
}
\value{
a named list containing information about the resampling that was carried out
}
\description{
Parses the id created by resampling
}

