% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{ps2003}
\alias{ps2003}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("ps2003")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 112 rows and 12 columns.
}
\source{
\url{https://eml.berkeley.edu/~saez/}
}
\usage{
ps2003
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("ps2003")}
}
\details{
Top fractile shares of income excluding capital gains,
1913-2024 (Table A1 of the latest "TabFig" vintage on Saez's site):
cumulative shares \code{p90_100} through \code{p99_99_100} and within-bracket
shares \code{p90_95} through \code{p99_9_99_99}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("ps2003")}
}
\keyword{datasets}
