% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{psz2018}
\alias{psz2018}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("psz2018")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 102 rows and 10 columns.
}
\source{
\url{https://gabriel-zucman.eu/usdina/}
}
\usage{
psz2018
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("psz2018")}
}
\details{
Equal-split pre-tax and post-tax national income shares of the
bottom 50\%, middle 40\%, top 10\% and top 1\% (plus pre-tax top 0.1\%),
1913-2014. \code{pretax_bottom50}, \code{pretax_middle40} and all \verb{posttax_*}
series start in 1962, when individualized micro data begin.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("psz2018")}
}
\keyword{datasets}
