% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{r2011}
\alias{r2011}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("r2011")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 280 rows and 23 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
r2011
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("r2011")}
}
\details{
Quarterly 1939-2008: present discounted value of defense news
(\code{pdvmil}, and \code{pdvmily} as a share of lagged GDP), real GDP and
components, government and defense spending, hours, population, real
manufacturing wage, CPI, BAA yield, T-bill and the average marginal
tax rate.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("r2011")}
}
\keyword{datasets}
