% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{r2016_govt}
\alias{r2016_govt}
\title{Ramey (2016) government spending panel}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 275 rows and 36 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
r2016_govt
}
\description{
Ramey (2016) government spending panel
}
\details{
Quarterly US panel behind Section 4 (government spending) of the
Handbook chapter: nominal NIPA components and deflators, hours,
population, Ramey military news and Ben Zeev-Pappa defense news.
Column definitions: \code{subset(variables, dataset == "r2016_govt")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("r2016")}
}
\keyword{datasets}
