% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{r2016_tax}
\alias{r2016_tax}
\title{Ramey (2016) tax shocks panel}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 283 rows and 32 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
r2016_tax
}
\description{
Ramey (2016) tax shocks panel
}
\details{
Quarterly US panel behind Section 4 (tax shocks) of the Handbook
chapter: real NIPA aggregates, tax receipts, Romer-Romer and
Mertens-Ravn narrative tax shocks. Column definitions:
\code{subset(variables, dataset == "r2016_tax")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("r2016")}
}
\keyword{datasets}
