% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{r2016_tech}
\alias{r2016_tech}
\title{Ramey (2016) technology shocks panel}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 275 rows and 46 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
r2016_tech
}
\description{
Ramey (2016) technology shocks panel
}
\details{
Quarterly US panel behind Section 5 (technology shocks) of the
Handbook chapter: Fernald TFP series, business output and hours, and
identified technology/news shocks from several papers. Column
definitions: \code{subset(variables, dataset == "r2016_tech")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("r2016")}
}
\keyword{datasets}
