% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resampleFileName}
\alias{resampleFileName}
\title{Calculate a file name for given parameters}
\usage{
resampleFileName(fun, countryAbbrev, formula, nest = NULL,
  nestStr = nestString(nest), sep = "_")
}
\arguments{
  \item{fun}{the function used for fitting}

  \item{countryAbbrev}{the country being fitted}

  \item{formula}{the formula used for the fitting}

  \item{nest}{if used, the nest employed for this fit. A 2-
  or 3-vector of integers.}

  \item{nestStr}{if used, the string for the nesting}

  \item{n}{the number of resamples being attempted}

  \item{sep}{the separator used to create the id string.
  Default is " : ".}
}
\value{
a string to be used as the id for this resample
}
\description{
Calculate a file name for given parameters
}

