% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{resampleFilePath}
\alias{resampleFilePath}
\title{Path to resample coefficients for the given parameters}
\usage{
resampleFilePath(prefix, countryAbbrev, fitfun, formula, nest = NULL,
  sep = "_", resamplePath, extension = "rds")
}
\arguments{
\item{prefix}{a prefix for the file name. Should not include the separator.}

\item{countryAbbrev}{the country being fitted}

\item{fitfun}{the function used for fitting}

\item{formula}{the formula used for the fitting}

\item{nest}{if used, the nest employed for this fit. A vector of 2 or 3 integers.}

\item{sep}{the separator used to create the id string. Default is "_".}

\item{resamplePath}{the path to the directory containing the resample data.
This is likely to be something like "data_resample/Calvin"}

\item{extension}{the filename extension (leading "." will be supplied if not present).}
}
\value{
a string representing the file name for these resample coefficients.
}
\description{
Path to resample coefficients for the given parameters
}
