% Generated by roxygen2 (4.0.0): do not edit by hand
\name{resampleModelsFileName}
\alias{resampleModelsFileName}
\title{Calculate a resample models file name for the given parameters}
\usage{
resampleModelsFileName(fun, countryAbbrev, formula, nest = NULL, sep = "_")
}
\arguments{
  \item{fun}{the function used for fitting}

  \item{countryAbbrev}{the country being fitted}

  \item{formula}{the formula used for the fitting}

  \item{nest}{if used, the nest employed for this fit. A
  vector of 2 or 3 integers.}

  \item{sep}{the separator used to create the id string.
  Default is "_".}
}
\value{
a string representing the file name for these resample coefficients.
}
\description{
Calculate a resample models file name for the given parameters
}

