% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-finance.R
\docType{data}
\name{rr2017}
\alias{rr2017}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("rr2017")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3060 rows and 3 columns.
}
\source{
\url{https://eml.berkeley.edu/~dromer/}
}
\usage{
rr2017
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("rr2017")}
}
\details{
Semiannual financial-distress score (0-14, higher means more
distress) for 30 OECD countries, 1967:1-2017:2. Long tibble: \code{date}
(first day of the half-year), \code{country}, \code{distress}. Author's site
extension of the original 24-country 1967-2012 series through 2017 and
6 more countries.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("rr2017")}
}
\keyword{datasets}
