% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rz2018}
\alias{rz2018}
\title{Ramey and Zubairy (2018) US historical fiscal data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 564 rows and 28 columns.
}
\source{
\url{https://econweb.ucsd.edu/~vramey/research.html}
}
\usage{
rz2018
}
\description{
Ramey and Zubairy (2018) US historical fiscal data
}
\details{
Quarterly US data 1889-2015 (population from 1875) used for the
state-dependent government spending multipliers: purchases, GDP,
deflator, unemployment, military news, T-bill, ZLB dummy, potential GDP,
debt and receipts. Column definitions:
\code{subset(variables, dataset == "rz2018")}.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("rz2018")}
}
\keyword{datasets}
