% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{s2015}
\alias{s2015}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("s2015")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1833 rows and 19 columns.
}
\source{
\url{http://www.econ.yale.edu/~shiller/data.htm}
}
\usage{
s2015
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("s2015")}
}
\details{
Monthly S&P Composite price, dividends, earnings, CPI and long
rate from 1871, the real series, CAPE, total-return CAPE, excess CAPE
yield and 10-year forward returns, as on Shiller's website.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("s2015")}
}
\keyword{datasets}
