% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-updated.R
\docType{data}
\name{sp500}
\alias{sp500}
\title{Stock market data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1780 rows and 4 columns.}
\usage{
sp500
}
\description{
adsfasdf
}
\keyword{datasets}
