\docType{data}
\name{State_distance}
\alias{State_distance}
\title{Haversine Distance between center of US states}
\format{numeric matrix}
\source{
  \url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\description{
  contains a matrix with great-circle distance in km
  between the center of US states.
}
\author{
  Florian Oswald
}
\references{
  \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}
\keyword{datasets}

