% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_distance.r
\docType{data}
\name{State_distMat}
\alias{State_distMat}
\title{Haversine Distance Matrix between center of US states}
\format{numeric matrix}
\source{
\url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\usage{
data(State_distMat)
}
\description{
contains a matrix with great-circle distance
in km between the center of US states.
}
\author{
Florian Oswald
}
\references{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

