\docType{data}
\name{State_distTable}
\alias{State_distTable}
\title{Haversine Distance Table between centers of US states}
\format{long table}
\source{
  \url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\description{
  contains a long table with great-circle distance in km
  between the center of US states.
}
\author{
  Florian Oswald
}
\references{
  \url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}
\keyword{datasets}

