% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-development.R
\docType{data}
\name{sz2016}
\alias{sz2016}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("sz2016")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 8 columns.
}
\source{
\url{https://gabriel-zucman.eu/uswealth/}
}
\usage{
sz2016
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("sz2016")}
}
\details{
Top wealth shares from capitalizing income tax returns,
1913-2012 (the Figure 1 series): \code{bottom90}, \code{top10}, \code{top5}, \code{top1},
\code{top05}, \code{top01} and \code{top001} (top 0.01\%) shares of household wealth.
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("sz2016")}
}
\keyword{datasets}
