% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_states.r
\docType{data}
\name{US_states}
\alias{US_states}
\title{Crosswalk of US State FIPS, Census Region and Division with PSID codes}
\format{long data.table}
\source{
\url{http://www.epa.gov/enviro/html/codes/state.html}, \url{http://www.census.gov/popest/about/geo/codes.html}
}
\usage{
data(US_states)
}
\description{
produces a data.table with state names, FIPS 
abbreviations. Also includes census region and division and PSID state code.

you will most likely want to merge this table into some 
dataset you come across and which is missing full state names,
or abbreviations, or FIPS code etc.
}
\author{
Florian Oswald
}

