% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_CPIHOSSL.r
\docType{data}
\name{USSTHPI}
\alias{USSTHPI}
\title{All Transactions house price index for the USA}
\format{monthly xts series}
\source{
\url{https://fred.stlouisfed.org/series/USSTHPI}
}
\usage{
data(USSTHPI)
}
\description{
U.S. Federal Housing Finance Agency, All-Transactions House Price Index for the United States [USSTHPI], retrieved from FRED, Federal Reserve Bank of St. Louis; \url{https://fred.stlouisfed.org/series/USSTHPI}, February 14, 2017.
index 1982-84=100.
}
\author{
Florian Oswald
}

