% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{wx2016}
\alias{wx2016}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("wx2016")}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 746 rows and 3 columns.
}
\source{
\url{https://www.atlantafed.org/cqer/research/wu-xia-shadow-federal-funds-rate}
}
\usage{
wx2016
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("wx2016")}
}
\details{
Monthly effective federal funds rate (\code{ff}) and the Wu-Xia shadow
rate (\code{shadow}), both on the last business day of the month, 1960-2022.
The shadow rate is only defined at the zero lower bound (NA otherwise).
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("wx2016")}
}
\keyword{datasets}
