% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{bbe2005}
\alias{bbe2005}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_title("bbe2005")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 511 rows and 121 columns.}
\source{
\Sexpr[stage=render,results=rd]{econdata:::bib_url("bbe2005")}
}
\usage{
bbe2005
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bbe2005")}
}
\section{Author(s)}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_author("bbe2005")}
}

\section{Year}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_year("bbe2005")}
}

\section{Journal}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_journal("bbe2005")}
}

\section{Doi}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_doi("bbe2005")}
}

\keyword{datasets}
