% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bbe2005}
\alias{bbe2005}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("bbe2005")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 511 rows and 121 columns.}
\usage{
bbe2005
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bbe2005")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("bbe2005")}
}
\keyword{datasets}
