% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{bq1989}
\alias{bq1989}
\title{Blanchard & Quah's "The Dynamic Effects of Aggregate Demand and Supply Disturbances"  (1989, AER)'
\itemize{
}}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\source{
value
}
\usage{
bq1989
}
\description{
Blanchard & Quah's "The Dynamic Effects of Aggregate Demand and Supply Disturbances"  (1989, AER)'
\itemize{
}
}
\keyword{datasets}
