% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bq1989}
\alias{bq1989}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("bq1989")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 159 rows and 3 columns.}
\usage{
bq1989
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("bq1989")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("bq1989")}
}
\keyword{datasets}
