% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_cpiausl.r, R/doc_CPIHOSSL.r
\docType{data}
\name{CPIAUCSL}
\alias{CPIAUCSL}
\title{CPI all urban consumers seasonally adjusted}
\format{monthly xts series}
\source{
\url{http://research.stlouisfed.org/fred2/series/CPIAUCSL}

\url{http://research.stlouisfed.org/fred2/series/CPIAUCSL}
}
\usage{
data(CPIAUCSL)

data(CPIAUCSL)
}
\description{
monthly CPI data for all urban consumers,
seasonally adjusted.
index 1982-84=100.

monthly CPI data for all urban consumers,
all items, seasonally adjusted.
index 1982-84=100.
}
\author{
Florian Oswald

Florian Oswald
}

