% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econdata-package.R
\docType{package}
\name{econdata-package}
\alias{econdata}
\alias{econdata-package}
\title{Robust Econometric Inference}
\description{
Provide data from past economic papers
}
\author{
\strong{Maintainer}: Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}

}
\keyword{internal}
