\docType{package}
\name{EconData}
\alias{EconData}
\alias{EconData-package}
\title{EconData}
\description{
  This package ships with a suite of datasets that I needed
  in economic research over the years. All of them are
  avaiable in the public domain, but none of them in an
  easy to use format. I provide the raw data as well as the
  functions that convert the data into a useable R dataset.
}
\author{
  Florian Oswald <Florian.oswald@gmail.com>
}

