% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econdata-package.R
\docType{package}
\name{econdata}
\alias{econdata}
\alias{econdata-package}
\title{econdata: Curated List of Economic Datasets}
\description{
The goal of {econodata} is to provide a curated list of datasets 
    that are used in the most influential economic papers.
}
\author{
\strong{Maintainer}: Kostas Vasilopoulos \email{k.vasilopoulo@gmail.com}

}
\keyword{internal}
