% Generated by roxygen2 (4.0.0): do not edit by hand
\name{energyTypeFromFormula}
\alias{energyTypeFromFormula}
\title{Extracts an energyType from a formula}
\usage{
energyTypeFromFormula(formula)
}
\arguments{
  \item{formula}{the formula that was fitted}
}
\value{
a string representing the energyType that was used for the fitting. NA if the sfModel was used.
}
\description{
Extracts an energyType from a formula
}

