% Generated by roxygen2 (4.0.0): do not edit by hand
\name{factorFromFormula}
\alias{factorFromFormula}
\title{Extracts a factor from a formula}
\usage{
factorFromFormula(formula)
}
\arguments{
  \item{formula}{the formula that was fitted}
}
\value{
a string representing the factor that was used for the fitting.
NA if the sfModel was not used.
}
\description{
Extracts a factor from a formula
}

