% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_FHFA_msa.r
\docType{data}
\name{FHFA_msa50}
\alias{FHFA_msa50}
\title{FHFA house price indices for 50 biggest MSAs over time}
\format{long}
\source{
\url{http://www.fhfa.gov/Default.aspx?Page=87}, \url{http://www.fhfa.gov/webfiles/26097/4q13hpicbsa_expandeddata.txt}
}
\usage{
data(FHFA_msa50)
}
\description{
contains 2 lists \code{qtr} and \code{yr},
which stands for yearly and quarterly frequency.
all data is in long format as data.tables.
}
\author{
Florian Oswald
}

