% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-updated.R
\docType{data}
\name{hprice}
\alias{hprice}
\title{Historical housing market}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\source{
value
}
\usage{
hprice
}
\description{
\itemize{
}
}
\keyword{datasets}
