% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kl2017}
\alias{kl2017}
\title{Killian Lutkepohl 2017 Book}
\format{US quarterly data from 1954-Q4 to 2007-Q4}
\source{
value
}
\usage{
kl2017
}
\description{
Killian Lutkepohl 2017 Book
}
\keyword{datasets}
