% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{kl2017}
\alias{kl2017}
\title{Killian Lutkepohl 2017 Book}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 213 rows and 4 columns.}
\usage{
kl2017
}
\description{
adsfsadf
}
\keyword{datasets}
