% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadCESResampleData}
\alias{loadCESResampleData}
\title{Parses the id created by resampling}
\usage{
loadCESResampleData(nest, energyType = "none", archive = NULL, baseResample)
}
\arguments{
  \item{id}{an \code{id} as a string}

  \item{sep}{the separator used to create the id string.
  Default is " : ".}

  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ".}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}

  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ".}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}

  \item{prefix}{the prefix for the model file names.}

  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ".}

  \item{factor}{the name of the factor as it appeared in
  the original data file. For example, "iK".}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}

  \item{modelType}{one of \code{"sf"}, \code{"cd"},
  \code{"cde"}, \code{"cese-(kl)"}, \code{"cese-(kl)e"},
  \code{"cese-(le)k"}, \code{"cese-(kl)e"}, or
  \code{"linex"}.}

  \item{countryAbbrev,}{a character string naming the
  country, if you want to use original data.}

  \item{baseResample}{the relative path of the top-level
  directory containing the resample data.}

  \item{nest}{the desired nest. One of \code{"(kl)"},
  \code{"(kl)e"}, \code{"(le)k"}, or \code{"(ek)l"}. If you
  specify nest=\code{"all"}, you'll get data for all nests.
  You'll need to specify energyType if you use
  nest=\code{"all"}.}

  \item{energyType}{the name of the energy type as it
  appeared in the original data file. For example, "iQ". If
  the energyType argument is missing or NA, you'll get data
  for the CES model without energy.}

  \item{archive}{path to a zip file archive}

  \item{baseResample}{the parent directory of all resample
  data, usually \code{"data_resample"}.}
}
\value{
a named list containing information about the resampling that was carried out
Path to resample data file

This function returns a string representing the relative file path for the resample data being requested.

the relative path of the file containing the data for the requested resample data.
Path to resample models file

This function returns a string representing the relative file path for the resample models being requested.

the relative path of the file containing the data for the requested resample data.
Generates paths to resample coefficients or models in String format

This function returns a string representing the relative file path for the resample coefficients or models being requested.

the relative path of the file containing the data for the requested resample data.
Directory for resample data

This function returns a string representing the relative directory path containing the resample data being requested.

the relative path to the directory containing the data for the requested resample data.
Loads and binds data for a CES resample ternary plot.

a data frame containing CES resample data for the given arguments.
}
\description{
Parses the id created by resampling
}

