% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/loaders.R
\name{loadPostProcessedData}
\alias{loadPostProcessedData}
\title{Loads post-processed data files}
\usage{
loadPostProcessedData(Source, kind = c("coeffs", "models", "fitted"),
  dir = file.path("data_postprocessed"))
}
\arguments{
\item{Source}{the data source to be loaded}

\item{kind}{one of \code{coeffs} (to load file named <Source>_Coeffs),
\code{models} (to load file named <Source>_Models), or
\code{fitted} (to load file named <Source>_Fitted).}

\item{dir}{specifies the directory from which to load the files.
Default is \code{data_postprocessed}.}
}
\description{
Loads post-processed data files
}

