% Generated by roxygen2 (4.0.0): do not edit by hand
\name{loadResampledData}
\alias{loadResampledData}
\title{Loads resampled coefficients from files}
\usage{
loadResampledData(path = "", archive = NULL, country = NULL,
  model = NULL, factors = NULL, sep = "_", prefix = c("coeffs",
  "models"))
}
\arguments{
  \item{path}{the directory containing the resampled
  coefficient files. Do not add a trailing file separator.}

  \item{archive}{a zip archive containing resampled
  coefficient files}

  \item{country}{a country or vector of countries.}

  \item{model}{a model or vector of models.}

  \item{factors}{strings of factors used in fitting the
  models.}

  \item{sep}{the separator within the file names between
  country, model, and factors}

  \item{prefix}{a character string naming prefix for file
  name.}
}
\value{
a string representing the energyType that was used for the fitting. NA if the sfModel was used.
}
\description{
Loads resampled coefficients from files
}
\details{
Supply either \code{path} or \code{archive} arguments.
Defaults are correct when only one is supplied, so long as archives are constructed with
all resample files at the root level of the zip archive.
}

