% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{loadResampledData}
\alias{loadResampledData}
\title{Loads resampled coefficients, models, or fits from files}
\usage{
loadResampledData(path = "", archive = NULL, country = NULL,
  model = NULL, factors = NULL, sep = "_", kind = c("coeffs", "models",
  "fitted"))
}
\arguments{
\item{path}{the directory containing the resampled coefficient files. Do not add a trailing file separator.}

\item{archive}{a zip archive containing resampled coefficient files}

\item{country}{a country or vector of countries.}

\item{model}{a model or vector of models of the form \code{c("sf", "cd", "ces", "linex")}, for example.}

\item{factors}{strings of factors used in fitting the models 
of the form \code{"iK+iL+iQp"} or \code{"iQp+iK+iL"}.}

\item{sep}{the separator within the file names between country, model, and factors}

\item{kind}{a character string that identifies which kind of resampled data to load, one of 
\code{coeffs}, \code{models}, or \code{fitted}.}
}
\value{
a data frame containing resampled coefficients (if \code{kind="coeffs"}),
a list of models which contain the resample fits (if \code{kind="models"}), or
a data frame containing resampled fits to the response variable (if \code{kind="fitted"}).
}
\description{
Loads resampled coefficients, models, or fits from files
}
\details{
Supply either \code{path} or \code{archive} arguments. 
Defaults are correct when only one is supplied, so long as archives are constructed with 
all resample files at the root level of the zip archive.
\code{kind} can be one of 
\code{"coeffs"}, to load resampled coefficients; 
\code{"models"}, to load models generated by the resample fits; or
\code{"fitted"}, to load fits to resampled data.
Models need to be generated with \code{save.data=TRUE} so that the original data can
be extracted from the model objects.
}

