\name{makeAbbreviations}
\alias{makeAbbreviations}
\title{make a US State <-> Abbreviation data.table}
\usage{
  makeAbbreviations(url = "http://www.epa.gov/enviro/html/codes/state.html")
}
\arguments{
  \item{url}{}
}
\value{
  NULL. saves data to packageroot/data
}
\description{
  make a US State <-> Abbreviation data.table
}
\seealso{
  Other makers: \code{\link{makeAllData}},
  \code{\link{makeMedianIncome}}
}

