% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeAllData}
\alias{makeAllData}
\title{make all datasets}
\usage{
makeAllData()
}
\description{
This builds all datasets in this package from its raw source. Most of the data come installed with package as excel or csv files. Some are downloaded from online sources.
}
\seealso{
Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeCPIUS}}, \code{\link{makeFHFA}},
  \code{\link{makeLincolnHomeValues}},
  \code{\link{makeMORTGAGE30US}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

