% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeBEAincome}
\alias{makeBEAincome}
\title{Per Capita annual personal income and per capita disposable income by state for 1949-now}
\usage{
makeBEAincome(root = "~/git/EconData")
}
\description{
data obtained from query to \url{http://bea.gov/iTable}, and
\url{http://bea.gov/iTable/iTable.cfm?reqid=70&step=1&isuri=1&acrdn=5}
}

