% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeInterstateMig}
\alias{makeInterstateMig}
\title{make Interstate Migration Transition Matrix}
\usage{
makeInterstateMig(root = "~/git/EconData")
}
\arguments{
\item{root}{name of package root directory}
}
\description{
make Interstate Migration Transition Matrix
}

