\name{makeLincolnHomeValues}
\alias{makeLincolnHomeValues}
\title{make Lincoln Institute Home and Land Value Dataset}
\usage{
  makeLincolnHomeValues(root = "~/git/EconData")
}
\arguments{
  \item{root}{path/to/your/package/root}
}
\value{
  NULL.
}
\description{
  reads data from the Lincoln Institute xls file.
}
\references{
  \url{https://www.lincolninst.edu/subcenters/land-values/}
}
\seealso{
  Other makers: \code{\link{makeAbbreviations}},
  \code{\link{makeAllData}}, \code{\link{makeCPIUS}},
  \code{\link{makeFHFA}}, \code{\link{makeMORTGAGE30US}},
  \code{\link{makeMedianIncome}},
  \code{\link{makeUS_coordinates}},
  \code{\link{makeUS_distance}}
}

