% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makers.r
\name{makeOwnershipRates}
\alias{makeOwnershipRates}
\title{make Homeownership rates by state over time}
\usage{
makeOwnershipRates(root = "~/git/EconData")
}
\description{
make Homeownership rates by state over time
}

