% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_Migration.r
\docType{data}
\name{Migration}
\alias{Migration}
\title{Migratory Flows between US states in 2012}
\format{long data.table}
\source{
\url{http://www.census.gov/hhes/migration/data/acs/state-to-state.html}
}
\usage{
data(Migration)
}
\description{
Census ACS estimate of the flow of migrants
from one US state to another for 2012.
}
\author{
Florian Oswald
}

