% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nber_rec}
\alias{nber_rec}
\title{NBER recession peak, trhough and duration}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 3 columns.}
\usage{
nber_rec
}
\description{
NBER recession peak, trhough and duration
}
\keyword{datasets}
