% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-updated.R
\docType{data}
\name{nber_rec}
\alias{nber_rec}
\title{NBER recession peak, trhough and duration}
\format{monthly data from 1854}
\source{
https://www.nber.org/cycles/NBER%20chronology.xlsx
}
\usage{
nber_rec
}
\description{
\itemize{
}
}
\keyword{datasets}
