% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{oil}
\alias{oil}
\title{Killian Lutkepohl 2017 Book - Oil VAR data}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 162 rows and 4 columns.}
\usage{
oil
}
\description{
fasdasd
}
\keyword{datasets}
