% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-books.R
\docType{data}
\name{oil}
\alias{oil}
\title{Killian Lutkepohl 2017 Book - Oil VAR data}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\source{
value
}
\usage{
oil
}
\description{
\itemize{
}
}
\keyword{datasets}
