\docType{data}
\name{Ownership}
\alias{Ownership}
\title{Home Ownership rates for US states over 2005-2012}
\format{long data.table}
\source{
  \url{https://www.census.gov/housing/hvs/data/rates/}
}
\description{
  quarterly US Census estimates of homeownership rates by
  state from 2005 to 2012.
}
\author{
  Florian Oswald
}
\keyword{datasets}

