% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{papers}
\alias{papers}
\title{List of the contained papers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 16 columns.}
\usage{
papers
}
\description{
The list contains all data sourced from the bibfile.
}
\keyword{datasets}
