% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{psy2015}
\alias{psy2015}
\title{Testing for Multiple Bubbles: Historical Episodes of Exuberance and Collapse in the S&P 500}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\source{
value
}
\usage{
psy2015
}
\description{
\itemize{
}
}
\keyword{datasets}
