% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{psy2015}
\alias{psy2015}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("psy2015")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 547 rows and 5 columns.}
\usage{
psy2015
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("psy2015")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("psy2015")}
}
\keyword{datasets}
