% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constants.R
\name{relevelFactor}
\alias{relevelFactor}
\title{Relevels a factor based on given levels}
\usage{
relevelFactor(x, levs)
}
\arguments{
\item{x}{a factor to be releveled}

\item{levs}{desired order for the levels of factor \code{x}.}
}
\value{
an object containing same data as \code{x} whose levels have been reordered according to \code{levs}.
}
\description{
Relevels a factor based on given levels
}
\details{
The length of \code{levs} may be longer or shorter than the number of levels in \code{x}.
If lengths are different, only \code{levs} that are also factors of \code{x}
are applied. 
Levels of factor \code{x} that do not appear in \code{levs} are ignored;
they will appear \emph{after} \code{levs} in the levels of the object returned from this function.
If none of \code{levs} are in the levels of \code{x}, an unmodified version of \code{x}
will be returned.
}
