% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_distTab.r
\docType{data}
\name{State_distTable}
\alias{State_distTable}
\title{Haversine Distance Table between centers of US states}
\format{long table}
\source{
\url{http://staff.washington.edu/glynn/dist_matrix.pdf}
}
\usage{
data(State_distTable)
}
\description{
contains a long table with great-circle distance
in km between the center of US states.
}
\author{
Florian Oswald
}
\references{
\url{http://www.r-bloggers.com/great-circle-distance-calculations-in-r/}
}

