% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{sw2001}
\alias{sw2001}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("sw2001")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 164 rows and 4 columns.}
\usage{
sw2001
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("sw2001")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("sw2001")}
}
\keyword{datasets}
