% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{sw2001}
\alias{sw2001}
\title{Stock & Watson (2001). "Vector Autoregressions," Journal of Economic Perspectives}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\usage{
sw2001
}
\description{
Stock & Watson (2001). "Vector Autoregressions," Journal of Economic Perspectives
}
\section{BLa}{


\itemize{
\item date:
\item infl:
\item un:
\item ff:
}
}

\keyword{datasets}
