% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets-papers.R
\docType{data}
\name{sw2001}
\alias{sw2001}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_title("sw2001")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 164 rows and 4 columns.}
\source{
\Sexpr[stage=render,results=rd]{econdata:::bib_url("sw2001")}
}
\usage{
sw2001
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("sw2001")}
}
\section{Author(s)}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_author("sw2001")}
}

\section{Year}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_year("sw2001")}
}

\section{Journal}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_journal("sw2001")}
}

\section{Doi}{
 \Sexpr[stage=render,results=rd]{econdata:::bib_doi("sw2001")}
}

\keyword{datasets}
