% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{u2005}
\alias{u2005}
\title{Uhlig's "What are the effects of monetary policy on output?" (2005, JME)}
\format{US quarterly data from 1960-Q1 to 2000-Q4}
\source{
value
}
\usage{
u2005
}
\description{
\itemize{
}
}
\keyword{datasets}
