% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{u2005}
\alias{u2005}
\title{\Sexpr[stage=render,results=rd]{econdata:::bib_authoryear("u2005")}}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 468 rows and 7 columns.}
\usage{
u2005
}
\description{
\Sexpr[stage=render,results=rd]{econdata:::bib_title("u2005")}
}
\references{
\Sexpr[stage=render,results=rd]{econdata:::bib_ref("u2005")}
}
\keyword{datasets}
