% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/doc_states.r
\docType{data}
\name{US_states}
\alias{US_states}
\title{Correspondance of US State with FIPS code, Census Region and Division}
\format{long data.table}
\source{
\url{http://www.epa.gov/enviro/html/codes/state.html}, \url{http://www.census.gov/popest/about/geo/codes.html}
}
\usage{
data(US_states)
}
\description{
produces a data.table with state names and FIPS
abbreviations. Also includes census region and division.

you will most likely want to merge this table into some
dataset you come across and which is missing full state names,
or abbreviations, or FIPS code etc.
}
\author{
Florian Oswald
}

